Update "Contributors" link in site-footer.html
The "Hugo Authors" link to Contributors is out of sync with the one found in the "Releases" page, for example. Both links currently resolve to the same destination.

If this continues to be the case going forward, then it is not a big deal. However, it is always best to keep links consistent to avoid possible broken links in the future.

The Release version is the most up-to-date as it is also used in other places in GitHub. The older link may become deprecated in future.