Fix search box text being invisible in Safari
Explicitly declare a transparent background color in the inline `style` declaration so the search box doesn't render white text on a white background in Safari. Remove `.bg-transparent` from the class attribute, since it doesn't exist in the built CSS file.

Explanation: for some reason (Purgecss misfire?) the CSS class `bg-transparent` does not appear in the built CSS file, so the assignment of the class to the `input` element is meaningless. The inline style declared using the `background` shortcut declaration lacks a background-color, which Chrome and Firefox interpret as a transparent background but which Safari just ignores, falling back to its default white background. This means text typed into the search field renders as white on white, effectively invisible. By explicitly declaring a transparent background color, the search box renders correctly in all three browsers (white text on black background).