Revise support for glossary of terms
In #267 we introduced a hack where, with type set to "glossary" in
front matter, level 3 headings followed by one or more paragraphs have
the same appearance as a description list. Each term has an id so it can
be used as a URL fragment.

This reverts those changes, but applies the behavior to all level 6
headings regardless of content type.

As of now, the only place we use level 6 headings is in the glossary.